pyproject.toml
[tool.poetry]
name = "aiven"
version = "0.1.0"
description = ""
authors = ["Petr Šebek <petrsebek1@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"
requests = "^2.27.1"
pytest = "^7.1.2"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
responses = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
